ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
pydantic==2.5.0
redis==5.0.1
httpx==0.25.2
python-multipart==0.0.6
aiofiles==23.2.1
bcrypt==4.0.1
